// test/src/pages/sandbox/users/ichibasan/sandbox-builder/types.ts
export type LayoutKey =
  | "outerSidebarStart"
  | "outerSidebarStartHeader"
  | "outerSidebarStartFooter"
  | "outerSidebarEnd"
  | "outerSidebarEndHeader"
  | "outerSidebarEndFooter"
  | "globalHeader"
  | "globalFooter"
  | "contentHeader"
  | "contentFooter"
  | "paneStart"
  | "paneStartHeader"
  | "paneStartFooter"
  | "paneEnd"
  | "paneEndHeader"
  | "paneEndFooter"
  | "innerSidebarStart"
  | "innerSidebarEnd";

export type ContentArea =
  | "globalHeader"
  | "globalFooter"
  | "contentHeader"
  | "contentBody"
  | "contentFooter"
  | "paneStartHeader"
  | "paneStartBody"
  | "paneStartFooter"
  | "paneEndHeader"
  | "paneEndBody"
  | "paneEndFooter"
  | "outerSidebarStartHeader"
  | "outerSidebarStartBody"
  | "outerSidebarStartFooter"
  | "outerSidebarEndHeader"
  | "outerSidebarEndBody"
  | "outerSidebarEndFooter"
  | "innerSidebarStart"
  | "innerSidebarEnd";